chore: move normalizeConfigFile to own folder
